Merge to main

+++ mkimage.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set -e
 
 # Supported machine name:
 #	sifive_unmatched
@@ -6,12 +7,8 @@
 #	starfive_jh7110
 # PS: "" means preserve variant
 
-# TODO:
-# jh7110: can't read uEnv.txt
-# d1: 
-# unmatched: redo partition for special uboot partition
-
-MACHINE=sifive_unmatched
+# MACHINE=sifive_unmatched
+machine_list=("sifive_unmatched" "allwinner_d1" "starfive_jh7110")
 BOARD=
 IMAGE_SIZE=4G
 IMAGE_FILE=""
@@ -25,24 +22,27 @@ CHROOT_TARGET=rootfs
 
 install_depends()
 {
-	echo "TODO: check if depends has been installed"
-	#apt install -y mmdebstrap debian-ports-archive-keyring \
-	#	qemu binfmt-support qemu-user-static curl wget \
-	#	libncurses-dev gawk flex bison openssl libssl-dev device-tree-compiler \
-	#	swig 
+	sudo apt install -y git gdisk dosfstools mmdebstrap qemu-system-misc binfmt-support qemu-user-static debian-ports-archive-keyring \
+		wget make gcc flex bison gcc-riscv64-linux-gnu libssl-dev bc dpkg-dev rsync 
 }
 
 unmount_image()
 {
 	echo "Finished and cleaning..."
-	umount -l "$CHROOT_TARGET"
-	losetup -d "$LOOP_DEVICE"
+	if mount | grep "$CHROOT_TARGET" > /dev/null; then
+		umount -l "$CHROOT_TARGET"
+	fi
+	if losetup -l | grep "$LOOP_DEVICE" > /dev/null; then
+		losetup -d "$LOOP_DEVICE"
+	fi
 	if [ "$(ls -A $CHROOT_TARGET)" ]; then
 		echo "folder not empty! umount may fail!"
 		exit 2
 	else
 		echo "Deleting chroot temp folder..."
-		rmdir "$CHROOT_TARGET"
+		if [ -d "$CHROOT_TARGET" ]; then
+			rmdir "$CHROOT_TARGET"
+		fi
 		echo "Done."
 	fi
 }
@@ -50,20 +50,34 @@ unmount_image()
 cleanup_env()
 {
 	echo "Cleanup..."
-	rm -rv kernel
-	rm -rv opensbi
-	rm -rv u-boot
-	rm -v *.deb
-	rm -v *.buildinfo
-	rm -v *.changes
-	rm -v *.bin
-	rm -v *.itb
+	if [ -d kernel ]; then
+		rm -rv kernel
+		rm -v *.deb
+		rm -v *.buildinfo
+		rm -v *.changes
+	fi
+
+	if [ -d opensbi ]; then
+		rm -rv opensbi
+	fi
+
+	if [ -d u-boot ]; then
+		rm -rv u-boot
+	fi
+
+	if [ -f *.bin ]; then
+		rm -v *.bin
+	fi
+
+	if [ -f *.itb ]; then
+		rm -v *.itb
+	fi
+
 	echo "Done."
 }
 
 main()
 {
-	# TODO: exception handling
 	install_depends
 	make_imagefile
 	pre_chroot
@@ -71,10 +85,7 @@ main()
 	make_kernel
 	make_bootable
 	after_chroot
-	# for debug:
-	chroot "$CHROOT_TARGET" bash
-	unmount_image
-	cleanup_env
+	exit
 }
 
 
@@ -89,17 +100,39 @@ source $(pwd)/scripts/pre_chroot.sh
 # Check root privileges:
 if (( $EUID != 0 )); then
     echo "Please run as root"
-    exit
+    exit 1
 fi
 
-# TODO: clean other things
-trap clean_env INT
-clean_env()
+if [ -z "$MACHINE" ]; then
+	echo "MACHINE not set!!"
+    exit 1
+else
+	if [[ " ${machine_list[*]} " =~ " ${MACHINE} " ]]; then
+    	echo "MACHINE=$MACHINE"
+	else
+		echo "$MACHINE is not compatible with this script!!"
+		exit 1
+	fi
+fi
+
+trap return 2 INT
+trap clean_on_exit EXIT
+
+clean_on_exit()
 {
-	echo "Ctrl+C exit!!"
-	unmount_image
-	rm -v "$IMAGE_FILE"
-	exit
+	if [ $? -eq 0 ]; then
+		unmount_image
+		cleanup_env
+		echo "exit."
+	else
+		unmount_image
+		cleanup_env
+		if [ -f $IMAGE_FILE ]; then
+			echo "delete image $IMAGE_FILE ..."
+			rm -v "$IMAGE_FILE"
+		fi
+		echo "interrupted exit."
+	fi
 }
 
 main
